+++ AWG/core/src/faces/awesome/model/PlayerCharacter.java
@@ -50,7 +50,7 @@ public class PlayerCharacter extends Character {
 
 
     //TODO lägga till facing
-    public void move(float dx, float dy) {
+    public void move(int dx, int dy) {
 
         //Kolla först om den tile man ska till är solid eller inte (här kommer det även kollas
         //om man tile är t.ex en dörr eller liknande som man kan gå in genom)
